move client assets to public directory
