+++ public/scripts/anti-cheat-system.js
