+++ public/scripts/bingo-anticheat-integration.js
