+++ public/scripts/bingo.js
