+++ public/scripts/firebase-leaderboard.js
