+++ public/scripts/polls.js
