+++ public/scripts/storage.js
