+++ public/scripts/username-validator.js
