+++ public/scripts/validation-analytics.js
