+++ public/scripts/verse.js
